Changed importer hashing of string input ids
A simple case as importing a node with input id = "1" and then importing
a relationship referring to a bad node with id = "" would connect the
relationship to the node with id = "1". The way the hashing function was
written any one character input id would match the empty input string.

This does NOT solve the general problem with false positives during
import when using input ids that has not been mapped but it removes this
incredibly simple case and simplifies the hashing function.

Cherry-picks