5.0 clear query caches fabric ()
* dbms.clearQueryCaches should also clear Fabric query cache for that database

* Reformat with Spotless

* Remove defensive check on FabricExecutor

FabricExecutor doesn't need a defensive null check because it always exists. If fabric is disabled then it still exists but the database's BootstrapService will simply not use it.

* Fix off by one bug

* Improve implementation of Fabric.clearCache

The implementation shall now clear the cache via mutable methods. This improves the implementation because the underlying cache is a Java mutable data structure, and by working directly with it we do not need to convert it to an immutable Scala data structure.

* FabricCache test coverage improved

- We removed a test in FabricExecutor which added no value
- We renamed a test and added the following tests in FabricPlanner
  - cache clears a context
  - cache clears only the given context
  - cache hits before being cleared, and misses after being cleared
  - cache returns the number of evictions

Co-authored-by: Mark Dixon <1756429+mnd999@users.noreply.github.com>