Checksum of multi versioned pages ()
* Checksum of multi versioned pages

This PR introduces the possibility of performing checksums over pages in multi versioned stores together with validation of those stored checksums on the next page load from persistent storage.

The checksum of the page will be calculated every time page is swapped out to an external storage device or when it's written out as part of the checkpoint. In case of different checksum for a particular page match exception is thrown and that particular page is not readable anymore. Access to any other page is still possible and the database is still operational.
The only way to restore that unreadable page is to restore it from a valid consistent backup.

* Review comments: allow some parts of swapper tests to be executed with enabled checksum (those that do not play with different page sizes), * instead of shift; remove incorrect ``-` in vector read

* Format code

Co-authored-by: Agent Smith <103414894+neo4jagentsmith[bot]@users.noreply.github.com>