Don't do extra validation of constraint indexes
After the constraint index has been created an exclusive
lock on the label is taken while validating the uniqueness
of the index. That lock is held until the constraint is committed.

Since shared lock on the label is taken on
setProp/deleteProp/addLabel/removeLabel it is blocked
from conflicting.

Extra verification of constraint index on constraint transaction commit should not be necessary.