5.0 Backup from multiple uris ()
You can now provide multiple addresses as the 'from' argument to a backup command as a comma separated list. Each address will be tried in order until one can provide a backup of the required database. If you specify a pattern as part of the target database, all servers will be contacted at the start to get the full list of matching databases, before backing up each found database in turn, as described above. If any server cannot be contacted, a warning will be raised and the command will exit with a non-zero code, although it will not interrupt the rest of the planned backup attempts.

This only applies to the 'new' backup command, currently called backup-next.