IO statistic per checkpoint ()
This PR changes the way how internal events are organized to allow collect statistics per flush, or group of flushes (checkpoint, etc). Statistic still propagated to top-level in the end as before but will allow having additional details to be visible and observable.

As a result of those changes, the checkpoint message will now include the number of flushed pages and how much is to the total number of pages; the number of neo4j io's completed and the used io limiter value.
For example:
  - `Checkpoint flushed 73 pages (7% of total available pages), in 46 IOs. Checkpoint performed with IO limit: unlimited.`
  - `Checkpoint flushed 600 pages (10% of total available pages), in 500 IOs. Checkpoint performed with IO limit: 800.`

Also, all of those except the percentage of pages exposed as additional checkpoint metrics and can be consumed by monitoring tools.