Setting namespaces validation  ()
* Setting namespaces validation

Before this PR it was possible to name your setting with any possible name. That was causing problems with understanding the setting scope and area of influence.
This PR restricts possible root namespaces for settings to be dbms, db, browser, server or internal. And it will be impossible to use settings with non-qualified names when strict validation will be enabled.
Since it's possible that some 3rd parties named their settings differently check will not be enforced for now when strict validation is disabled.

* Add more temporary namespaces and some test settings rename