Add flag to import tool for strict node id check
While importing relationships and referring to node by string id to a
node that has not been imported there is a potential risk (very low)
that that id results in the same hash as an existing node and the
relationship would be imported succesfully but will refer to the wrong
node.

If relationships in the CSV files might refer to non-existing nodes then
this flag should be set. Setting this flag impacts performance since two
lookups will be needed for each imported relationship.

This flag is turned on by default which will make imports using string
ids slower. Turn off the strict flag to get back to former performance
if that is considered safe.