Flush log file channel after header write ()
Before this PR it was possible to observe files without headers that are present for some time and may cause some exceptions during file reads. While most of the places that do file log reading will retry or be ready for the file to be in this state it produces additional stack traces and causes some retries and can be avoided easily.