Makes RelationshipIndexCursor reading from store explicit
To cater for the scenario where only the information already
available from the index is enough. This fixes an issue where
the store read was lazy on first data accessor after a successful
next, where if the relationship got concurrently deleted it could
not be read and instead return "-1" data or fail.

Also removes superfluous companion RelationshipScanCursor
with relationship index cursors because it's enough with the
index cursor for all use cases.