Stop echoing passwords in Cypher Shell when output is redirected
A user can redirect output and still use the interactive prompt to supply the
password, for example `cypher-shell -u neo4j -f statements.cypher > out.txt`.
Previously, masking the password didn't work, and it leaked into the
terminal (either we used jline incorrectly or it is a bug in jline). With this
change we use tty and stty, if available, to disable echoing.

We still leak passwords in Windows, and always have. But I consider this a
separate issue, since it's not part of the regression.