Add PEM parser that do not use bouncy castle
Bouncy castle adds a lot of overhead in terms of meta-space consumption. It supports, for us, an unreasonable amount of formats/ciphers/algorithms. This completely removes bouncy castle from compile scope so it will not be shipped with the final installation.

This change also introduces a lot of tests to verify that we can actually parse different formats. The tests will be used as a reference when creating the accompanying documentation so that we can explicitly state what we support.