page cursor refactoring ()
This refactoring makes more parts of MuninnPageCursor immutable. As cursors are always tied to the same PagedFile we can remember all the related fields and don't need to re-initialise them.
Cursors now explicitly throw IllegalStateException on pin if cursor was closed already, instead of NPEs from null pagedFile reference.
The only place where cursor can be reused after close - cached instance of linked cursor.

rewind method is removed as it not used in prod code now and it's implementations in different cursors looked quite broken.

Closing of paged file now reacts to thread interrupts. This makes it possible to terminate failing tests that stuck because someone failed to release page locks.
Interrupt handling:

flushAndForceInternal in paged file checks thread interrupted flag in busy loop. if thread is interrupted it will limit the number of attempts to grab a page lock, flush whatever possible and throw InterruptedException at the end
when it is called from close(unmap) that is not part of checkpoint, InterruptedException is swallowed and interrupted status of the thread is reset, allowing to flush other files if needed
when it is called as part of checkpoint, interrupted status of the thread is reset and UnderlyingStorageException is thrown, so checkpoint fails