Frees uncommitted IDs in tx commit/rollback
Previously node/relationship IDs (which are allocated while building
the transaction state) were not freed in commit/rollback and
instead automatically marked as deleted later on. Those IDs couldn't
be reused within the same session, but needed a restart of the db
to be reused. Now they can be reused shortly after being "unallocated"
as part of tx commit/rollback.