Freki implementation of allRelationshipsScan
This implementation is node based as compared to Record format since
that makes most sense for Freki. But this means that size of batches
will be unpredicatable since it will vary a lot.

Might make sense to make the calculation of batch sizes use number of
nodes for Freki instead of as now number of relationhips.