Don't allow creating node key constraints that data doesn't conform to
When we create index-backed constraints we release the label lock during index population and then
take it again when population has finished and validate that the index is actually unique
(nothing conflicting happened during population).
But for node key constraints (property existence + uniqueness) it is not enough, uniqueness
will hold but not existence. We only checked the property existence before we created
the index. At each committing transaction property existence is checked against all constraints.
But while we are populating the index backing nodeKey the committing transactions can't see
the nodeKey constraint yet, so they can create nodes that don't uphold the constraint, and
then that was never checked. This means we could end up with a constraint that our data
didn't actually conform to.