Add somewhat 'standard' *nix exit codes
Other than 0 for success and non-zero for error, there is no formal
standard for exit codes.

It's common for 1 to mean "general" error, but anything specific is
usually left to the developer.

sysexits is the closest to something, somewhat standard, and was
introducted into BSD, and used by many *nix tools to indicate a few
more specific errors, other than 1.

This adds a class that wraps the constants from sysexits.h; though it
would be ideal to be an enum, their use in annotations prevents this.