Use ExitCode where otherwise hardcoded for consistency
Initially toyed with AbstractCommand::execute returning an int; however,
as commands are written to use CommandFailedException to control flow;
it would require a large refactor to be consistent.

Therefore, sticking to current exception-flow, but using a more
centralised set of exit codes.

Explicitly use USAGE and SOFTWARE exit codes for AbstractCommand, when
incorrect usage and unhandled exceptions occur.