5.0 cluster platform extension ()
Cluster API for testing 5.0 cluster.

Currently waits for DEFAULT database at startup. This is not inteded in the future when allocator defaults to equal counts.

There is currently a bug in routing for read replicas where they cannot route to a writer for a database it is not hosting. Therefor a lot of tests currently run without secondaries