Find/Select StorageEngineFactory without PageCache
We only check existence of files to select storage engine and never open any files to check. Maybe this was the intention once upon a time but can we settle with something simpler? This way, it's easier to find which engine you operate under, just knowing FileSystem and a layout. Makes it easier to adapt things for freki