Don't fail transactions that was started before constraint was committed ()
* Don't fail transactions that was started before constraint was committed

Back in the days there was this thing called HA (high availability). It was a poor man cluster solution and it worked... if you squinted a lot and kept your distance.

Anyway, in this cluster solution it was possible to write on slaves (followers). And hence this validation was needed in order to: "To protect from race conditions between constraints coming online on master long before they are enforced on slaves, master now rejects transactions that were started before a constraint was brought online."

Now we instead grab label write lock on the affected label which solves the problem completely without needing to fail any transactions.

* Remove test case that verifies old behaviour