Undirected relationship scans should only find self-loops once
For queries with `(a)-[r]-(b)` we might either scan for one of the nodes, say a, and find r and b or we might solve it
with a relationship scan and from there find a, r, and b in one go. However, for a self-loop (`CREATE (a)-[:LOOP]->(a)`),
these are not equivalent operation if the undirected relationship scan emits two rows for each relationship it finds also
when the relationship is a self-loop.