Add enable server procedure ()
* Add enable server procedure

This adds the start of the enable server backend and temporary procedure (until the surface exists).
This procedure is hard to test because servers are currently auto-enabled, and there
is no way to make sure we can enable the server using the new method instead of the old one.
Auto-enabling servers is still present and will be modified in a different PR, at which point
it might be easier to write an IT that enables a new server using the new procedure.

* Perform allocations when enabling server

* Add empty string check

* Simplify code

* Clarify prefix based behaviour