Avoid lots of file system lookups while looking for existing file profiles ()
* Avoid lots of file system lookups while looking for existing file profiles.

Before this change set, we were spending a lot of time in `sameRelativePath` check.
That was because instead of pointing to profiles as much as possible directly we tried to look up and find possible matches. In addition to that those matches were not that effective. Those are not a problem on a small number of databases, but on envs with several hundreds of those, it becomes a bit more expensive.

* Add test to page cache warmer to make sure we can locate existing profiles for files, even if they have same file name but reside in different directories.

* Review comments: additional test cases