Freki: consistency checks around the big value property store ()
Consistency checks for the big property store

A new flag has also been introduced to the ConsistencyFlags checkPropertyValues that is disabled by default. Any inconsistencies discovered when this flag is enabled will be flagged as warnings.

A visitRecordChain has also been introduced that will scan through the store and uses the above flag to check the stored properties against what we expect to be present in the node/relationship stores.