Page cache file-backed page versions storage. ()
* Page cache file-backed page versions storage.

Before this PR experimental multi-versioned format was operating on top of the in-memory area where all old page versions resided.
The memory area was size limited and when it was full it was game over for that multi-versioned run.

This PR removes that temporal in-memory implementation of the version storage and replaces it with
page cache backed. Each database has its own version storage that is
cleaned up on restart. Database store files share the same version storage.
Any file page mutation creates a copy in a version store that is organized in chains.
A version of a store introduced in this PR does not know how to remove old snapshots while the database is running.

* Move multi versioned record format and page file backed version storage to enterprise.

* Review comments: do not close read cursor, javadoc, constants move