Trigram index filter out false positives and support range queries
Most queries against trigram index need to be filtered for false positives. This happens in the DefaultNodeValueIndexCursor and DefaultRelationshipValueIndexCursor.

For trigram index to be able to fully replace text index it needs to support RANGE queries. This is done by doing an inclusive range scan over the first trigram (or ngram if search terms are shorter) of 'from' and 'to' and then filter out false positives through the store.