Use plain incremented atomic long for generated tx ids in bolt server. ()
It was observed during the investigative runs that we spend a lot of time in locks related to bolt transaction id generation. Before this PR it was using a random UUID for that. It looks like we do not need those ids to be that complex and this PR switching them to be strings that are generated based on incremented atomic long.