Add DatabaseFile marker for existence
Add method to check the existence of the "marker" file from the layout;
to avoid manual checking of the existence of the metadata store which
does not have a common location accross formats.

This manual checking is done in a few tests.

Currently keeping the marker as their respective metadata stores, though
could easily be changed to a completely independent file in future; thus
allowing the metadata store itself to be used in tests.

Treating this marker as a database file for consistency, and semantics;
as all users of the streams of files gets put in a Set; it doesn't
matter if the marker file, is the same as another file.

Co-authored-by: Matthew Parnell <matthew.parnell@neo4j.com>