Move CreateDbTool to neo4j-tools ()
* Move CreateDbTool to neo4j-tools

Some things that needed to change:
- DbStatistics moved to neo4j-commons because it was needed by both CreateDbTool and ZippedStore implementations
- MigrateDbSetup uses DatabaseStartupAwaitingListener to wait for dbs to come online before returning because test-dbms class is not accessible from new module.
- DbCreator uses Preconditions instead of AssertJ to assert consistency check was successful

* Make sure to open correct FileSystem when creating path to ZippedStore .zip file

And move implementation of common methods of ZippedStoreCommunity and ZippedStoreEnterprise to ZippedStore interface.

* Append more exception information for debugging purposes

* ZippedStore create base uri for file system creation recursively

When trying to get reference a .zip file among resources by path we might need to init the corresponding FileSystem. In order to do this we need to extract the correct uri schema. This is now done recursively to make it possible to run both when running tests locally and in CI.

* Copy zipped store to a new directory before use instead of trying to get direct path

Getting the path to the zipped file among resources is incredibly difficult so instead we stream the content to a known place and use that zip instead.