Migrate stores without any written transactions present during the whole 4.x series ()
Enhance migration to support scenarios when legacy databases transaction logs do not have any transactions in logs in new formats. And instead, those logs were never updated ever after the update to 4.x version.
To be able to support that as part of this PR ability to parse inlined checkpoints was restored (both 3.5 and 4.x formats).