Stateless CSV value extractors
The previously state-full design was very annoying and error-prone
because most extractors were used by multiple threads and had
to be maintained and cloned to work in a multi-threaded setting.
Besides, the small benefit that the state-less design had was
only to avoid some unboxing... an advantage that no actual
main code was using.

One side-effect of this is to fix the Ldbc imports which had
a very rare race which made some CSV header parsing fail unexpectedly,
e.g. exposed by EmbeddedCoreDense1SnbInteractiveExecutionTest