Fix buffering issue with InputStream from DefaultFileSystemAbstraction
The len argument to InputStream::read is the maximum amount of data that
can be read; whereas it seems to have been treated as the b.length, the
length buffer into which the data should be read.

This causes an issue when this
DefaultFileSystemAbstraction.NativeByteBufferInputStream is used
within a BufferedOutputStream, as is effectively done within
DefaultFileSystemAbstraction::openAsInputStream, that the offset is
subtracted from the length twise, thus can result in a negative amount
of data to read; and throws accordingly.