Improve memory tracking in ForsetiLockManager in presence of memory limits ()
This PR changes memory tracking in ForsetiClient so it allocates memory before trying to take a lock, and then doing it's best to release it if lock isn't taken. This is done to prevent incorrect negative memory calculation when hitting memory limit during lock acquisition.

There are still cases when it's possible that allocated memory will not be immediately released in case of failed lock acquisition, i.e. because of deadlock. Such memory will be released anyway when the lock client is closed, so no extra handling is done for such cases.