Non-recursive unwrapping of partial predicates.
This is quicker.
The recursive rewriter was not needed since we should only end up with top-level PartialPredicates when doing cardinality estimation.

The only apparent exception would be when estimating cardinality for plans
from the OrLeafPlanner, but there we plan a `Filter` on all union branches
to cover the full predicate before joining them together.

E.g.

```
.union()
.|.filter("a.prop CONTAINS 'world") // solves Contains(Property(Variable(a),PropertyKeyName(prop)),StringLiteral(hello))
.|.nodeIndexOperator("a:A(prop)") // solves PartialPredicateWrapper(IsNotNull(Property(Variable(a),PropertyKeyName(prop))),Contains(Property(Variable(a),PropertyKeyName(prop)),StringLiteral(hello)))
```