Do not generate CartesianProducts and Joins multiple times per single pattern.
In SingleComponentPlanner.planSinglePattern we generated plans on top of each leaf.
But the CPs we generated did not depend on the actual leaf we were mapping at the moment.
And the joins only half depended on the actual leaf we were mapping at the moment.

This lead to generating the same CP / Join multiple times and then remove them right away
again, because we put everything into a Set.

Now we only generate the CP / Joins once.