Jvm additional setting should not split inside quoted values
The jvm additional parse did not correctly detect quoted values for
properties. Give `-Dfoo="with spaces"` would result in 2 arguments,
`["-Dfoo=\"with","spaces\""]`, which in turn would result in the command
line `'-Dfoo="with' 'spaces"'`. This change will correctly parse the
above string to a single argument `'-Dfoo="with spaces"'`.