Fix inconsistent index behaviour
For queries like `MATCH (n:L) WHERE n.prop={'key': 42} RETURN n.prop` we have inconsistent
behaviour depending on if we use an index or not. Without an index we will get no result whereas
when using an index we got an error saying that

>Property values can only be of primitive types or arrays thereof

This changes the behaviour so that we consistently behave the same way as when there is no index.

changelog: Don't throw when seeking the index for values that cannot be indexed.