Add --max-off-heap-memory and --threads to check command
Also unifies how tools that generally goes by the philosophy of
"the more memory the better and faster I'll run" calculates how
much memory to use. Tools like import, incremental import and
consistency checking.

--max-off-heap-memory can be supplied either with an actual
amount of memory, or a percentage of the currently free memory
on the machine.

This simplifies the consistency checker code, simplifies
the usage of the tool, makes the defaults more sane and allows
for better control of its memory usage as a whole.