Fix regression after removing drop result
In 4.3 we removed DropResults which we used for cases where we statically know that
a predicate is false, as in `MATCH (n) WHERE 1 < 0 RETURN n`. Instead of using DropResult we
inserted a Limit(0), however in pull-based runtimes when there is an eager operator below
limit we would still compute the result which means that these kinds of queries may run slower.

See https://github.com/neo4j/neo4j/issues/12926