Fixes bootloader Ctrl+C issue on Windows
Apparently having SIGINT/SIGTERM signal handlers on the parent
process interfered with those of the child process on Windows
such that if both were installed the shutdown hook that shuts down
the dbms weren't run on the child process at all.

Fix by not having those SIGINT/SIGTERM handlers on the parent process
on Windows.