Fix bug in AssertingMultiNodeIndexSeekOperator
Fix bug in pipelined runtime where some queries using AssertingMultiNodeIndexSeekOperator would fail incorrectly with `Merge did not find a matching node...`. For example in queries like `UNWIND [0,1] AS id MERGE (n:A:B {id: id})`, if there is a uniqueness constraint on the id property for both A and B labels.