Fix NaN range search with index use ()
All comparison with NaN results in FALSE.
- Exact match on NaN should be empty
- Any range predicate including NaN as a boundary should be empty
- However, all_entries scan and exists scan should still find NaN.

Implementation
For exact predicate on NaN, create specific IncomparableExactPredicate and for range predicate containing NaN, use IncomparableRangePredicate. Do instance of check in NativeIndexReader (superclass of RangeIndexReader).

An alternative implementation would be to hide the return types in PropertyIndexQuery and introduce a new EmptyPredicate with IndexQueryType.EMPTY or something like this and implement it as a valid predicate type without instance of check. But the predicate types have leaked a lot and is considered to big of an effort to change now. Might happen in the future.

Co-authored-by: Anton Persson <anton.persson@neotechnology.com>