Fix NPE in procedure-compiler for array types ()
Fix NPE in procedure-compiler for array types

.. that was reporter in Github issue #12936. The RecordFieldTypeVisitor
did not handle that array types are dispatched to visitArray. When this
occurred, the class falls back on the base class' defaultAction (return
null) which leads to the NullPointerException.

To avoid this type of error, we override the defaultAction to provide a
sensible default.

Finally, the commit also adds the specific type that fails validation,
to the error message to provide better context to the user.