Let CommandCreationContext be able to lock and use transaction seq nr
As preparation for Freki relationship ids the Freki implementation of
CommandCreationContext need to be able to take locks when relationship
ids are reserved.

The generation of ids also needs to know the current
transaction sequence number and the ability to on demand retrieve the
oldest active transaction sequence number to be able to reserve ids that
does not match with anything seen by a living transaction.