AST and Fabric disagree on order for Union queries
For queries like `RETURN 1 AS a, 2 AS b UNION RETURN 2 AS b, 1 AS a`, there is a
disagreement on whether we should return fields in order `a, b` or `b, a`. Fabric
defines takes the RHS as the order whereas the AST defines the order from the LHS.
This leads to columns being swapped at runtime.

changelog: Fix bug where columns could be swapped for UNION queries.