Fixes an issue scanning relationship batches
It seems the Freki implementation was under the impression that
the sizeHint parameter was something else. Corrected it and also
separated relationship and node scan batch code so that relationship
scan batches do some averaging of number of relationships per node
so that sizeHint can be "somewhat" followed. Also made some unit
tests in this area a bit less strict with asserting specific numbers.