Add serverID to grouping key for SHOW DATABASES ()
Add serverID to grouping key for SHOW DATABASES

When `server.bolt.advertised_address` is set on a cluster to override the endpoint of cluster members this can result in the grouping key for the aggregation that rolls up aliases for cluster members that are not primaries being the same.

The effect of this is the rows get compressed together, instead of two servers with ther own set of aliases, one row is produced with all the aliases. This can be fixed by adding the serverID into the grouping key.