Provide error for invalid record type in procedure-compiler ()
This is a paper-cut that was discovered when attempting to implement a
procedure without reading the documentation too closely.

When defining the return type `Stream<String>` the type exploration
would continue down to the private member variable `byte[] value` which
currently crashes. Once the crash is fixed the error message will still
be nonsensical.

Per the documentation:

    The record is defined per procedure, as a class with only public,
    non-final fields. The types, order and names of the fields in this
    class define the format of the returned records.

In order to provide the user with a more helpful error message this
commit compares the type `T` with the allowed types to see if the user
is doing the same mistake as I did.