Dev load csv lfi ()
* Add FileURLAccessRuleTest cases

Improve test coverage to include different inputs:

- non escaped
- single escaped
- double escaped
- triple escaped
- overlong escaped

Also introduce tests which highlight a bug in the FileURLAccessRule logic, by allowing paths to be traversed.

* Add LoadCsvAcceptanceTest cases

Added main test cases to this file. Coverage isn't as extensive as in FileUrlAccessTest because these integration tests are slow.

Also fixed some of the existing tests.

* Normalize URI in FileURLAccessRule

The path was being normalised but not the URI. The Java Path and URI classes originate from two different RFC documents which have different concerns and different normalisation algorithms. We need to normalise the URI before we construct a path with it, in order to protect the DBMS from potential vulnerabilities.

* Adapt tests to the Windows OS