TransactionImpl avoid closing kti if it has already been closed
When we execute additional Cypher statement inside transaction event listener ‘beforeCommit’ and it fails we will call KernelTransactionImplementation.close twice.
Inside commit we call transaction.execute which causes the transaction to rollback and close, and then the commit will finish by closing the transaction again. The result is that the transaction is returned to the txPool twice and the numberOfActiveTransactions counter is off by one.

This is solved by checking if kti has already been closed inside of TransactionImpl, before closing.