Update indexesBySchemaAndType key in SchemaCache to be custom type instead of Pair ()
It was noticed that as part of command generation we quite often visit that collection and some relatively significant amount of time is wasted by doing constant hashCode/equals checks. That was partly because of Pair usage instead of proper pojo, partially because evaluations were not memoized. This change set fixes both of those.