Remove FileUrlAccessRuleTest special cases
These special cases didn't need to be added to assert the safety of the latest security fix during which they were added. And unfortunately they are so fine-grained that they don't behave the same way on multiple operating-systems and/or jdks. Removing for now, we will decide later if we need to add them back.