Test incorrect tx metrics changed on read -> write
Active transaction metric lifecycle:

1. new tx ==> active_read++
2. tx state is created
      txState == null  -->  txState != null
        ==>   read tx  -->  write tx
   a. active_read--
   b. active_write++
3. tx close
   if  read tx ==> active_read--
   if write tx ==> active_write--

Semantically active_read and active_write should always be non-negative.

The upgrade from read to write tx in (2) was only on the existence of
transaction state; however, there was a semantically different
definition of a write tx, which also included if there were changes.

There can be situations where transaction state was created, and no
changes exist. Changes are based on revisions.

* In the same transaction
  create a node (rev++), then delete the same node (rev++)
  is considered a change (rev >= 2)

* Trying to delete a node, already deleted and committed in a concurrent
  transaction; creates state, but makes no changes.

In such cases, when closing, despite having transaction state; the
transaction is treated as a read tx, rather than a write. This has the
resulted of decrementing active_read, rather than active_write;
therefore, active_read is one fewer, and active_write to be one more.
This is of course accumulative.

This adds tests asserting that both example above should count as write
transactions; the latter fails.