Fix case where tx metrics active_read go negative
Remove the inconsistency when defining write transactions.

Either, it is a write transaction when there are changes, or if there is
state (regardless of changes).

Though at first, the former seems more semantic; after a discussion,
the latter was chosen. Not only was simplest, and how we generally treat
it; it carries the semantics that a write operation was attempted, and
thus should not be a read transaction.