GBPTree can be opened in read-only mode
For this to happen it needs to accept a readOnly flag and stop
using DatabaseReadOnlyChecker, and here's why:

DatabaseReadOnlyChecker is a high-level concept of a database
dynamically switching between read-only mode and writable mode.
The only reason DatabaseReadOnlyChecker got passed into the
StorageEngine and all its internals everywhere was that GBPTree
accepted that parameter, and the only reason GBPTree used it was
so that it could throw exception on trying to create the file
if it didn't exist.

Funnily making that decision based on DatabaseReadOnlyChecker
is wrong because that mode is volatile and can change from
one instant to the next, due to config changes.
Also, no other store does a check like this.

Instead, GBPTree now accepts a readOnly boolean which has an
additional function, in addition to preventing file creation if missing,
and that is to prevent any change to the file, including:

- automatic bumping of generation at instantiation
- writers
- checkpoints

all of which were previously done anyway even if the
DatabaseReadOnlyChecker turned out to be read-only at the
time of instantiation. This meant that opening e GBPTree with e.g.
readOnly() would still bump generation and force new state to the file.
Not anymore.