Remove fsync from GBPTree construction ()
* Remove fsync from GBPTree construction

Before, GBPTree would flush the file a few times during its first construction in order to have metadata, state pages and the first empty tree structure stable on disk. This is not needed because those things will be flushed anyway on the first checkpoint and if we never reach the first checkpoint, all data needs to be recovered. We only need to keep the data in the file if it has seen a checkpoint and we can identify if a checkpoint has ever happened by looking at the state pages. If both of them are empty, the tree has never seen a checkpoint and needs to be recreated. In that case we will overwrite the minimal tree pages with zeroes as preparation to start over.

Remove header writer from GBPTree constructor
Since we no longer flush the file during construction we don't write the header. Instead, the first header in the tree will come with the first checkpoint.
- Adapt a few GBPTreeTest to this behaviour
- Remove/Change use pattern of header writer from classes that use GBPTree

IndexedIdGenerator
Instead of reading idsPerEntry from the header, simply use configured value directly and fail if it doesn't match with metadata in index. In this way we can avoid reading the header before opening the tree.
Also fix a faulty if-statement that would not rebuild .id file if generation==STARTING_GENERATION.

TokenIndex and NativeIndex
We don't write the POPULATING byte during construction, but non-existing state pages header will also be treated as POPULATING by TokenIndexProvider / NativeIndexProvider

Test fixes
The two most common test fixes for the change are:
- Number of pins, unpins, hits and pageFaults has changed when starting up a new db and there are many tests that assert on exact numbers. Those numbers have been adjusted.
- Many GBPTree tests assumed that GBPTree could just be opened and then closed directly and then be read from disk on next open. This is not the case anymore, so a checkpoint has been added before close in those cases.

Visit tree with empty state pages
If the state pages are empty (or non existing) it means that the tree has not been checkpointed. This state will be observed when visiting the tree through #visit or #unsafe before the first checkpoint. In those cases the reported state will be `FIRST_TREE_STATE` which is a "fake" representation of how the disk state should have looked if it was flushed in constructor.